refactor: 겹쳐있는 디렉토리 수정 (#19)
