+++ src/main/java/project/backend/business/user/UserService.java
@@ -1,20 +1,19 @@
-package project.backend.business.user.service;
+package project.backend.business.user;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import project.backend.business.user.implement.UserReader;
 import project.backend.entity.user.User;
-import project.backend.repository.user.UserRepository;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class UserService {
 
-  private final UserRepository userRepository;
+  private final UserReader userReader;
 
   public User findUserById(Long id) {
-    return userRepository.findById(id)
-                         .orElseThrow(() -> new IllegalArgumentException("유저를 찾을 수 없습니다. id: " + id));
+    return userReader.findUserById(id);
   }
 }
