+++ src/main/java/project/backend/presentation/user/controller/UserController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import project.backend.business.user.service.UserService;
+import project.backend.business.user.UserService;
 import project.backend.common.auth.aop.AssignCurrentUserInfo;
 import project.backend.common.auth.aop.CurrentUserInfo;
 import project.backend.entity.user.User;
